backend/emprestou_backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-JWT-Extended
Flask-Cors
SQLAlchemy-JSON
Pillow

